LPS-56971 move preprocess config to query package
